Implemented bank form on manage and create

+++ WebApp/lucianna_finaltest_comp303w/src/components/ManageBankById.jsx
@@ -1,10 +1,9 @@
-import React, { useState } from 'react';
+import {useState} from 'react';
 import {deleteBankById, getBankById, updateBankById} from '../services/bankService.js';
 import '../css/App.css'
 import BankForm from "./BankForm.jsx";
 
 function ManageBankById() {
-    const [banks, setBanks] = useState([]);
     const [bankId, setBankId] = useState('');
     const [idToDelete, setIdToDelete] = useState('');
     const [bankName, setBankName] = useState('');
@@ -15,10 +14,28 @@ function ManageBankById() {
     const [bankATMs, setBankATMs] = useState('');
     const [errorMessage, setErrorMessage] = useState('');
 
+    const clearValues = () =>{
+        // Clear form on successful submission
+        setBankId('')
+        setBankName('');
+        setBankYear('');
+        setBankEmp('');
+        setBankAddress('');
+        setBankBranches('');
+        setBankATMs('');
+        setErrorMessage('');
+    }
+
     const handleSearchById = () => {
         getBankById(bankId)
             .then((response) => {
-                setBanks([response.data]);
+                setBankId(response.data.bankId)
+                setBankName(response.data.bankName);
+                setBankYear(response.data.bankYear);
+                setBankEmp(response.data.bankEmp);
+                setBankAddress(response.data.bankAddress);
+                setBankBranches(response.data.bankBranches);
+                setBankATMs(response.data.bankATMs);
             })
             .catch((error) => {
                 console.error('Error fetching bank by ID:', error);
@@ -27,17 +44,24 @@ function ManageBankById() {
     };
 
     const handleDeleteById = () => {
-        deleteBankById(idToDelete)
-            .then((response) => {
-                setBanks([response.data]);
-                alert('Bank deleted successfully');
-                setIdToDelete('')
-            })
-            .catch((error) => {
-                console.error('Error deleting bank by ID:', error);
-                setErrorMessage('Error deleting bank: ' + error.message);
-            });
-    }
+        const confirmDelete = window.confirm(`Are you sure you want to delete the bank with ID ${idToDelete}?`);
+        if (confirmDelete) {
+            deleteBankById(idToDelete)
+                .then(() => {
+                    alert('Bank deleted successfully');
+                    setIdToDelete('');
+                    clearValues()
+                })
+                .catch((error) => {
+                    console.error('Error deleting bank by ID:', error);
+                    setErrorMessage('Error deleting bank: ' + error.message);
+                });
+        } else {
+            console.log('Deletion canceled by the user.');
+            setIdToDelete('');
+            clearValues()
+        }
+    };
 
     const handleSubmit = (e) => {
         e.preventDefault();
@@ -59,62 +83,64 @@ function ManageBankById() {
         };
 
         updateBankById(bankId, updatedBank)
-            .then((response) => {
+            .then(() => {
                 alert('Bank updated successfully');
-                // Clear form on successful submission
-                setBankId('')
-                setBankName('');
-                setBankYear('');
-                setBankEmp('');
-                setBankAddress('');
-                setBankBranches('');
-                setBankATMs('');
-                setErrorMessage('');
+                clearValues()
             })
             .catch((error) => {
                 setErrorMessage('Error updating bank: ' + error.message);
             });
-
-
     };
 
     return (
         <div>
-            <h2>Manage Bank by ID</h2>
-
+            <h2>Edit Bank By Id</h2>
             <section>
                 {/*Search by id*/}
-                <div>
+                <div className="input-button-container">
                     <input
                         type="text"
                         placeholder="Search by Bank ID"
                         value={bankId}
                         onChange={(e) => setBankId(e.target.value)}
                     />
-                    <button onClick={handleSearchById}>Search by ID</button>
+                    <button onClick={handleSearchById}>Search</button>
                 </div>
             </section>
 
             <section>
+                {/*Form */}
+                <BankForm handleSubmit={handleSubmit}
+                          btnDisplay="Save changes"
+                          bankId={bankId}
+                          bankName={bankName}
+                          setBankName={setBankName}
+                          bankAddress={bankAddress}
+                          setBankAddress={setBankAddress}
+                          bankYear={bankYear}
+                          setBankYear={setBankYear}
+                          bankEmp={bankEmp}
+                          setBankEmp={setBankEmp}
+                          bankBranches={bankBranches}
+                          setBankBranches={setBankBranches}
+                          bankATMs={bankATMs}
+                          setBankATMs={setBankATMs}
+                          errorMessage={errorMessage}/>
+            </section>
+            <hr/>
+
+            <section>
                 {/*Delete by id*/}
-                <div>
+                <h2>Delete By Id</h2>
+                <div className="input-button-container">
                     <input
                         type="text"
                         placeholder="Delete by Bank ID"
                         value={idToDelete}
                         onChange={(e) => setIdToDelete(e.target.value)}
                     />
-                    <button className="btn-danger" onClick={handleDeleteById}>Delete by ID</button>
+                    <button className="btn-danger" onClick={handleDeleteById}>Delete</button>
                 </div>
-            </section>
-
-            <section>
-                {/*Form section*/}
-                {banks.map((bank) => (
-                    BankForm(["Edit Bank"], handleSubmit, ["Save changes"], bank, bankName,
-                        setBankName, bankAddress, setBankAddress, bankYear, setBankYear, bankEmp, setBankEmp,
-                        bankBranches, setBankBranches, bankATMs, setBankATMs, errorMessage)
-                ))}
             </section>
         </div>
     );
